retrieve CSS selector through CDP
